tubequery/requirements.txt
# ============================================================
# TubeQuery - Python Dependencies
# ============================================================

# Transcript fetching
youtube-transcript-api>=0.6.2,<1.0.0
yt-dlp>=2024.11.0

# Embeddings (runs locally)
sentence-transformers>=3.3.0,<4.0.0
torch>=2.5.0,<3.0.0

# Vector store (runs locally)
chromadb>=0.5.0,<1.0.0

# LLM API
google-genai>=1.0.0,<2.0.0

# Environment + utilities
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.0.0,<3.0.0

# UI
streamlit>=1.41.0,<2.0.0

# HTTP (for YouTube Data API calls)
httpx>=0.28.0,<1.0.0
